Related shows lookup and add when setting up a show. Added searching by genres to api functions. Cleanup

+++ data_access.py
@@ -10,16 +10,6 @@ def __get_table():
     return __get_database().Table('tvshows')
 
 
-def get_current_show(alexaid):
-    table = __get_table()
-    response = table.scan(
-        FilterExpression=Key('alexa_id').eq(alexaid) & Attr('watching').eq(True)
-    )
-    if response['Items']:
-        return response['Items'][0]
-    return None
-
-
 def get_all_shows(alexaid):
     table = __get_table()
     response = table.scan(
@@ -30,33 +20,18 @@ def get_all_shows(alexaid):
     return None
 
 
-def set_show_watch_status(alexaid, show_id, value):
-    table = __get_table()
-    response = table.update_item(
-        Key={
-            'alexa_id': alexaid,
-            'id': show_id
-        },
-        UpdateExpression="set watching = :w",
-        ExpressionAttributeValues={
-            ':w': value
-        },
-        ReturnValues="UPDATED_NEW"
-    )
-    return response
-
-
 def clear_table(alexaid):
     table = __get_table()
     shows = get_all_shows(alexaid)
-    with table.batch_writer() as batch:
-        for s in shows:
-            batch.delete_item(
-                Key={
-                    'alexa_id': s['alexa_id'],
-                    'id': s['id']
-                }
-            )
+    if shows:
+        with table.batch_writer() as batch:
+            for s in shows:
+                batch.delete_item(
+                    Key={
+                        'alexa_id': s['alexa_id'],
+                        'id': s['id']
+                    }
+                )
 
 
 def insert_show(alexaid, id, show_name):
@@ -123,4 +98,18 @@ def __insert_data():
     table.put_item(Item=dexter)
     table.put_item(Item=wire)
 
-
+# example on how to update
+# def set_show_watch_status(alexaid, show_id, value):
+#     table = __get_table()
+#     response = table.update_item(
+#         Key={
+#             'alexa_id': alexaid,
+#             'id': show_id
+#         },
+#         UpdateExpression="set watching = :w",
+#         ExpressionAttributeValues={
+#             ':w': value
+#         },
+#         ReturnValues="UPDATED_NEW"
+#     )
+#     return response
